API/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
tenacity
pytest
